販売手数料計算用のjsファイル作成し記述移動、全体jsファイルには作成したjsファイルを読みこむ記述追加、何の記述ファイルか明確にした
